requirements.txt
torch>=1.12.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0
biopython>=1.79
pyBigWig>=0.3.18
pyfaidx>=0.6.4
plotly>=5.3.0
jupyter>=1.0.0
h5py>=3.1.0
statsmodels>=0.13.0
scipy>=1.7.0
pyyaml>=6.0
argparse>=1.4.0
joblib>=1.1.0
captum>=0.5.0
shap>=0.40.0
optuna>=3.0.0